Fix excessive memory consumption for np.loadtxt (Work in Progress).
The fix consist of parsing the array first to estimate the number of
rows and then create directly an array instead of storing it in
intermediary python tuples.

Memory consumption is divised by 20 on a 11MB compressed dataset.

TODO: verify for structured arrays.
TODO: It could be done maybe more efficiently by fixing the block size
and then iterating over file.read(blocksize)?